Refactor: Split types.rs into focused modules
Split the monolithic types.rs file (7008 lines) into 11 focused modules
for better maintainability and code organization.

Changes:
- Created types/ directory with the following modules:
  * common.rs - Core types, SHType enum, ClonedVar, type aliases
  * mesh.rs - Mesh and MeshVar types
  * wire.rs - Wire, WireRef, WireState, enum/object helpers
  * shard_ref.rs - ShardRef and AutoShardRef
  * metadata.rs - Type metadata, ExposedTypes, From impls
  * refcounted.rs - RCObjectVar trait and RefCounted<T>
  * param.rs - ParamVar and ShardsVar
  * strings.rs - Strings utilities and ExposedTypesIterator
  * seq.rs - Sequence types and iterators
  * table.rs - Table types and iterators
  * mod.rs - Module declarations and re-exports

- Removed original types.rs
- Fixed all compilation errors and import issues
- Verified build succeeds with cargo check

Benefits:
- Easier navigation and code discovery
- Reduced cognitive load when working with type system
- Better separation of concerns
- Improved compilation times (modular compilation)
